[CHERIseed] Remove AtomicBool and use atomic_uint8_t instead
The extra abstraction introduced by AtomicBool can be confusing due to
the overloaded = operator. Use atomic_uint8_t directly with the helper
functions: atomic_load_relaxed and atomic_store_relaxed. Although this
makes the code more verbose, it is easier to understand. Besides, this
is more in line with what other sanitizers do.

+++ compiler-rt/lib/cheriseed/cheriseed_interface_internal.h
@@ -167,7 +167,7 @@ void __cheriseed_stack_cap_get(__cheriseed_cap_t *cap);
 // -------------------------------------
 
 SANITIZER_INTERFACE_ATTRIBUTE
-void __cheriseed_enable_cheri_semantics(int enable);
+void __cheriseed_enable_cheri_semantics(u8 enable);
 
 SANITIZER_INTERFACE_ATTRIBUTE
 __cheriseed_cap_t *__cheriseed_strerror(__cheriseed_cap_t *result, int code);
@@ -176,7 +176,7 @@ SANITIZER_INTERFACE_ATTRIBUTE
 int __cheriseed_set_signal_handle_mode(__cheriseed_cap_t *context, int mode);
 
 SANITIZER_INTERFACE_ATTRIBUTE
-void __cheriseed_enable_invoke_signal_handlers(int enable);
+void __cheriseed_enable_invoke_signal_handlers(u8 enable);
 
 SANITIZER_INTERFACE_ATTRIBUTE
 void __cheriseed_static_init(void);
